app/requirements.txt
customtkinter
pillow
aiohttp
pydantic
SQLAlchemy
pydantic-settings
redis
fastapi